code/backend/utils/passwordutil/password_bcrypt.go
package passwordutil

import (
	"golang.org/x/crypto/bcrypt"
)

// EncryptPassword 对密码进行加密
func EncryptPassword(password string) (string, error) {

	hashedPassword, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
	if err != nil {
		return "", err
	}
	return string(hashedPassword), nil
}

// ComparePasswords 比较密码和加密后的密码是否匹配
func ComparePasswords(password, hashedPassword string) bool {
	err := bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
	return err == nil
}
